Users can now provide contact names when sending messages
When a user does something like `h!send "Dennis Leaman" This is a test message.` the app will search for a mobile number associated with Dennis Leaman and send it to the first number it finds. If none is found, it returns a blank String which will need be handled by Hermes replying with an error.